# src/main/java/org/sohan/BookManagementSystem/Auth/AuthenticationResponse.java
package org.sohan.BookManagementSystem.Auth;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationResponse {
    public String token;
}